Consolidate recipe blueprint reqs
Rather than having three separate members, use a single build_reqs
member.
